+++ README.md
@@ -38,7 +38,7 @@ FE 레벨1 영화 리뷰 미션
 1. 배너에 첫 번째 인기 영화의 정보가 표시된다.
 2. 최초 진입 시 인기 영화 최대 20개가 표시된다.
 3. 더 보기 버튼을 누르면 최대 20개가 추가된다.
-4. 더 이상 보여줄 영화가 없으면 더 보기 버튼이 사라진다.
+4. 더 이상 보여                                            줄 영화가 없으면 더 보기 버튼이 사라진다.
 
 ### 검색 시
 
@@ -47,3 +47,6 @@ FE 레벨1 영화 리뷰 미션
 3. 더 이상 보여줄 영화가 없으면 더 보기 버튼이 사라진다.
 4. 검색 결과가 없으면 `검색 결과가 없습니다.`를 아이콘과 함께 표시된다.
 5. 로고를 누르면 메인 화면으로 돌아온다.
+
+
+
