+++ src/Logo.ts
@@ -1,0 +1,15 @@
+class Logo {
+  #logo;
+
+  constructor() {
+    this.#logo = document.querySelector(".logo");
+  }
+
+  bindEvent() {
+    this.#logo?.addEventListener("click", () => {
+      location.reload();
+    });
+  }
+}
+
+export default Logo;
