+++ src/store.ts
@@ -1,0 +1,6 @@
+const PageStore = {
+  popularMoviePage: 1,
+  searchMoviePage: 1,
+};
+
+export default PageStore;
